+++ src/main/java/io/github/ryuu/adventurecraft/gui/GuiTree.java
@@ -34,7 +34,7 @@ public class GuiTree extends Screen {
 
     @Override
     public void init() {
-        this.treeSize = new GuiSlider2(4, 4, 4, 10, String.format("Tree Size: %.2f", Float.valueOf(this.tree.size)), (this.tree.size - 0.5f) / 3.5f);
+        this.treeSize = new GuiSlider2(4, 4, 4, 10, String.format("Tree Size: %.2f", this.tree.size), (this.tree.size - 0.5f) / 3.5f);
         this.buttons.add(this.treeSize);
         this.prevValue = this.treeSize.sliderValue;
     }
@@ -48,7 +48,7 @@ public class GuiTree extends Screen {
         this.fill(0, 0, this.width, this.height, Integer.MIN_VALUE);
         if (this.prevValue != this.treeSize.sliderValue) {
             this.tree.size = this.treeSize.sliderValue * 3.5f + 0.5f;
-            this.treeSize.text = String.format("Tree Size: %.2f", Float.valueOf(this.tree.size));
+            this.treeSize.text = String.format("Tree Size: %.2f", this.tree.size);
             this.world.method_246(this.blockX, this.blockY, this.blockZ);
             this.world.getChunk(this.blockX, this.blockZ).method_885();
         }
